+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - In `typescript` config, setting `filesTypeAware: false` now fully disables the config for running type-aware rules. Same if an empty array is provided.
+- Added the ability to disable running type-aware rules (or even any TypeScript rules) on the files specified in `enforceTypescriptInScriptSection` option.
 
 ## 0.4.0